Now saves the scores of logged decks
need to add rollover buffering

+++ DeckClassifier/src/DeckScore.java
@@ -2,9 +2,9 @@
 public class DeckScore {
 
 	public String name;
-	public float score;
+	public int score;
 	
-	public DeckScore(String s, float f) {
+	public DeckScore(String s, int f) {
 		name = s;
 		score = f;
 	}
